+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.9.3b
-- Fixed starting trainer battles with a fainted lead pokemon. Battles now start with the first non-fainted pokemon in the player's party.
+- Fixed allowing starting a trainer battle with a fainted lead pokemon. Battles now start with the first non-fainted pokemon in the player's party.
 - Fixed changing a trainer's group leaving a duplicate file in its old group directory.
+- Fixed commands not being disabled when mod is in a corrupt state.
+- Added separate permission node "selfdot.trainers.battle" for command /trainers battle
 
 # 0.9.3
 - Fixed new trainers not initialising with default data, resulting in a corrupt trainer file.