+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 0.9.4
+- Added Generation 5 based trainer AI.
+
 # 0.9.3b
 - Fixed allowing starting a trainer battle with a fainted lead pokemon. Battles now start with the first non-fainted pokemon in the player's party.
 - Fixed changing a trainer's group leaving a duplicate file in its old group directory.